move mvt code into tests
